fix(mimir): address code review findings (NIU-670)
- LintBadge.tsx: add missing CSS import
- RoutingPage.tsx: replace bare Fragment with <Fragment key={rule.id}> to
  fix missing-key React warning; remove now-redundant key on inner <tr>
- format.ts: extract shared formatDuration + formatTimestamp(iso, dateStyle)
  from DreamsPage and RavnsPage, eliminating the copy-paste duplication;
  DreamsPage passes 'medium', RavnsPage uses the default 'short'
- LintPage.tsx: replace hardcoded MOCK_RAVNS with useRavns() so the assignee
  dropdown is driven by the service layer
